+++ legacy/python/platform/src/python_platform.rs
@@ -190,7 +190,7 @@ impl Platform for PythonPlatform {
                                         id: dep_project_id.to_owned(),
                                         scope: DependencyScope::Production,
                                         source: DependencySource::Implicit,
-                                        via: Some(dep_name.clone()),
+                                        via: Some(format!("package {dep_name}")),
                                     });
                                 }
                             }
@@ -438,7 +438,7 @@ impl Platform for PythonPlatform {
 
         command.with_console(self.console.clone());
         command.args(&task.args);
-        command.envs(&task.env);
+        command.envs_if_not_global(&task.env);
 
         if let Ok(python) = self.toolchain.get_for_version(&runtime.requirement) {
             if let Some(version) = get_proto_version_env(&python.tool) {